requirements.txt
Django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt